% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{checkMeta}
\alias{checkMeta}
\title{checkMeta}
\arguments{
\item{meta}{List; Length 3 list with the Wells, Columns, and Plate sheets,
respectivley.}
}
\value{
Nothing.
}
\description{
Checks that assumptions concerning the metadata in Google Drive are fufilled.
}
\author{
Jason Serviss
}
\keyword{internal}
