% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.datesFormat}
\alias{.datesFormat}
\title{.datesFormat}
\arguments{
\item{meta}{Output from \code{\link{getPlateMeta}} function.}
}
\value{
List of tibbles with ymd formatted dates.
}
\description{
Date formatting helper to \code{\link{getPlateMeta}}.
}
\author{
Jason Serviss
}
\keyword{internal}
