% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.layout}
\alias{.layout}
\title{.layout}
\arguments{
\item{format}{Numeric; Indicates which plate to setup. Can be 384 or 96.}
}
\value{
A tibble with the plate layout.
}
\description{
Helper for \code{\link{resolvePlateMeta}}. Sets up the plate layout.
}
\author{
Jason Serviss
}
\keyword{internal}
