% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.processKeys}
\alias{.processKeys}
\title{.processKeys}
\arguments{
\item{data}{A tibble with potentially dupicated keys in the colnames.}
}
\value{
A character vector of the duplicated keys without their suffix.
}
\description{
Helper for \code{\link{resolvePlateMeta}}.Extracts duplicated key names with
a ".x" or ".y" suffix.
}
\author{
Jason Serviss
}
\keyword{internal}
