% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.resolve}
\alias{.resolve}
\title{.resolve}
\arguments{
\item{data}{Tibble; including the keys to be resolved.}

\item{key}{Character; the key to be resolved. key.x and key.y are expected
in colnames(data).}
}
\value{
A character vector of the duplicated keys without their suffix.
}
\description{
Helper for \code{\link{resolvePlateMeta}}. Resolved key precedence.
Expects the greater suffix (i.e. .y > .x) to have greater precedence.
Expects only .x and .y suffixes.
}
\author{
Jason Serviss
}
\keyword{internal}
