% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.safe1}
\alias{.safe1}
\title{.safe1}
\arguments{
\item{plate}{Character; Corresponds to the plate to process metadata for.
Must correspond to a file name in the path argument.}

\item{path}{Character; The location of the plate metadata on Google Drive.}
}
\description{
Checks for file on googledrive before attempting to download.
}
\author{
Jason Serviss
}
\keyword{internal}
