% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{.safe2}
\alias{.safe2}
\title{.safe2}
\arguments{
\item{path}{Character; The location of the plate metadata locally.}
}
\description{
Checks that all sheets are present before loading the file.
}
\author{
Jason Serviss
}
\keyword{internal}
