% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngeMetadata-package.R
\docType{package}
\name{EngeMetadata}
\alias{EngeMetadata}
\alias{EngeMetadata-package}
\title{A package for analyzing tissue spatiality in scRNAseq data.}
\description{
Description
}
\details{
Handles R import of Google Drive based metadata for the Enge lab
}
\author{
Author: Jason T. Serviss
}
\keyword{package}
