% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{getPlateMeta}
\alias{getPlateMeta}
\title{getPlateMeta}
\arguments{
\item{plate}{Character; Corresponds to the plate to process metadata for.
Must correspond to a file name in the path argument.}

\item{path}{Character; The location of the plate metadata on Google Drive.}

\item{verbose}{Logical; Indicates if function should be verbose.}
}
\value{
List; Length 3 list with the Wells, Columns, and Plate sheets,
respectivley.
}
\description{
Downloads and reads the metadata into R.
}
\author{
Jason Serviss
}
\keyword{internal}
