% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{metadata}
\arguments{
\item{plate}{Character; Corresponds to the plate to process metadata for.
Must correspond to a file name in the path argument.}

\item{path}{Character; The location of the plate metadata on Google Drive.}
}
\value{
A tibble with the metadata specified in the file.
}
\description{
Downloads and processes a metadata, formated according to a standard
template, for a specific plate and path on Google Drive.
}
\examples{

\dontrun{metadata("test1", 'Enge_lab/GFP_mouse/Annotation/package_testing')}

}
\author{
Jason Serviss
}
