% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{resolvePlateMeta}
\alias{resolvePlateMeta}
\title{resolvePlateMeta}
\arguments{
\item{meta}{List; Length 3 list with the Wells, Columns, and Plate sheets,
respectivley. Typically from the \code{\link{getPlateMeta}} function.}
}
\value{
A tibble with the resolved metadata.
}
\description{
Resolves the precedence of the of the keys in the metadata file with the
hierarchy Wells > Columns > Plates.
}
\author{
Jason Serviss
}
\keyword{internal}
